requirements.txt
openai
botpy
python-pptx
weasyprint
reportlab
pypdf
pdf2image
pillow
jinja2
markdown
pyyaml
